Made the map zoom level logic smarter.
Made the map zoom level default to the level that includes a radius of 25 miles (on mobile devices) OR a radius that includes at least one meal location, whichever is greater.

The logic, however, is going to be problematic if/when we have thousands of locations in the DB.  It has to iterate through and calculate the distance from the user to each one.  Google maps doesn't (currently) have a way that makes that easier (from what I can tell from my googling).

One day, we'll probably have to make it so the app has the user's location BEFORE it will even retrieve a list of meal locations from the server.  We'll just grab them all for now, though.